Step 1: Rework the repo client-building machinery.
This one's a _doozy_. Rework the entire client-building machinery into
completely agnostic client interfaces, with all of the GitHub-specific
code & work contained in a single GitHubRepositoryClient object that
translates GH-specific API calls and objects into our generic objects.
This way, all of the OctoKit blast radius is there and in the client
factory, which has to create the actual GH client to pass to our client
wrapper.

Still needs a lot of work:

1. Implementing BitBucket.
2. Implementing an HttpClientFactory-like cache for client objects that
   reference the same API host. All IRepositoryClient instances
   referring to github.com should share the same client, for as long as
   we want it to live.
3. Maybe some fleshing out of the generic models that we have, and/or
   better conversion into immutable records...